+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added option to TimeRemainingColumn to estimate the remaining time using an Exponential Moving Average instead of averaging the recent past.
+
 ## [13.7.1] - 2024-02-28
 
 ### Fixed